\name{bkcde-package}
\alias{bkcde-package}
\alias{bkcde}
\docType{package}
\title{
\packageTitle{bkcde}
}
\description{
\packageDescription{bkcde}
}
\details{
%%  ~~ An overview of how to use the package, ~~
%%  ~~ including the most important functions ~~
}
\author{
\packageAuthor{bkcde}

Maintainer: \packageMaintainer{bkcde}
}
\references{
%%  ~~ Literature or other references for background information ~~
}
\keyword{package}

%% Uncomment below to imitate parts of library(help = bkcde)
%\section{The \file{DESCRIPTION} File}{\packageDESCRIPTION{bkcde}}
%\section{Documentation Index}{\packageIndices{bkcde}}

\seealso{
%%  ~~ Optional links to other man pages, e.g. ~~
%%  ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%%  ~~ Optional simple examples of the most important functions ~~
}
